requirements.txt
cn2an==0.5.23
fastapi==0.135.1
keybert==0.9.0
langchain_classic==1.0.3
langchain_community==0.4.1
langchain_core==1.2.19
langchain_milvus==0.3.3
langchain_neo4j==0.8.0
langchain_ollama==1.0.1
langchain_openai==1.1.11
langchain_text_splitters==1.1.1
langgraph==1.1.2
langsmith==0.7.20
matplotlib==3.8.0
neo4j==6.1.0
numpy==2.4.3
ollama==0.6.1
onnx==1.20.1
onnxruntime==1.20.1
optimum==2.1.0
psutil==5.9.0
pydantic==1.10.12
pymilvus==2.6.10
python-dotenv==1.2.2
pywin32==305.1
rake_nltk==1.0.6
redis==7.3.0
Requests==2.32.5
scikit_learn==1.2.2
seaborn==0.13.2
sentence_transformers==5.3.0
setuptools==80.9.0
setuptools==68.2.2
spacy==3.8.11
torch==2.10.0
torch_xla==2.9.0
tqdm==4.67.1
tqdm==4.65.0
transformers==5.3.0
uvicorn==0.42.0
wordcloud==1.9.6
